requirements.txt
Flask==2.3.3
Werkzeug==2.3.7
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.0.5
gunicorn==21.2.0
python-dotenv==1.0.0
Flask-Login==0.6.2
Flask-Migrate==4.0.4
psycopg2-binary==2.9.7
Flask-WTF==1.1.1
email-validator==2.0.0.post2 